refactor(web): move the shared row formatters to a React-free leaf module
custom-plan-diff.ts documents that it stays free of React/JSX so it can be
unit-tested in isolation; importing the shared formatters from plan-spec.ts
pulled React in transitively via plan-tier-meta.
