+++ clients/web/src/domains/settings/billing/plan-spec.test.ts
@@ -7,13 +7,7 @@ import {
   FREE_STORAGE_GIB,
 } from "@/domains/settings/billing/plan-tier-meta";
 
-import {
-  creditRowLabel,
-  machineLabel,
-  packageHighlights,
-  packageSpecs,
-  storageRowLabel,
-} from "./plan-spec";
+import { machineLabel, packageHighlights, packageSpecs } from "./plan-spec";
 
 describe("packageSpecs", () => {
   test("uses the free/base baseline for a null package", () => {
@@ -140,22 +134,6 @@ describe("packageHighlights", () => {
     expect(
       packageHighlights({ machine_size: "small" } as ProPackage),
     ).toHaveLength(3);
-  });
-});
-
-describe("storageRowLabel", () => {
-  test("spells out the storage row", () => {
-    expect(storageRowLabel(30)).toBe("30 GB storage");
-  });
-});
-
-describe("creditRowLabel", () => {
-  test("drops the cents on a whole-dollar amount", () => {
-    expect(creditRowLabel(45)).toBe("$45 of bundled credits");
-  });
-
-  test("keeps the cents on a sub-dollar amount", () => {
-    expect(creditRowLabel(0.5)).toBe("$0.50 of bundled credits");
   });
 });
